pyproject.toml
[project]
name = "django_project"
version = "0.1.0"
description = "django project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django==5.2.9",
    "psycopg2==2.9.11",
]

[tool.uv]
link-mode = "copy"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
mypy_path = "."

[[tool.mypy.overrides]]
module = ["debug_toolbar.*"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.ruff.lint]
ignore = [
]

select = [
   "I",
]

[dependency-groups]
dev = [
    "ruff==0.14.7",
    "pytest==9.0.1",
    "django-debug-toolbar==6.1.0",
    "pytest-django==4.11.1",
    "factory-boy==3.3.3",
    "djlint==1.36.4",
    "django-stubs[compatible-mypy]==5.2.8",
    "djcheckup>=0.5.0",
]


